requirement.txt
mysql-connector-python==8.0.33
streamlit==1.33.0
SQLAlchemy==2.0.29
